% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.r
\name{descj}
\alias{descj}
\title{JSON description}
\usage{
descj(dat)
}
\arguments{
\item{dat}{\code{list}. a JSON list.}
}
\value{
\code{list}. The result contains three parts. \code{key_summary} presents the description
of keys. \code{value_summary} presents the description of values. And \code{stream_summary}
presents the description of keys' direct upstream keys and downstream keys.
}
\description{
Give descriptive information about the JSON list, such as the key frequency,  the nesting information and the value distribution.
}
