% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_one.r
\name{diffj_one}
\alias{diffj_one}
\title{Single JSON Diff}
\usage{
diffj_one(json_new, json_old, sep = "@")
}
\arguments{
\item{json_new}{list. The new json list.}

\item{json_old}{list. The old json list.}

\item{sep}{character. A character/string used to separate full keys in the nesting path.
Defaults to @ to avoid the occasional overriding. Not allowed to use some risky words like . and \.
See \code{flattenj_one()}.}
}
\value{
list. The difference between two JSON objects. The \code{difference} stores the data frame about the details of each change.
And the \code{summary} provides some useful summarization.
}
\description{
calculating the difference between two JSON objects.
}
\details{
This function can find the different between two JSON/list objects.
The difference information is categorized into three types: key addition, key removal and value change.
All these differences can be reflected by '+(add)', and '-(delete)' chng_type in the new JSON.
}
\examples{
library(mojson)
j1 = list(a=list(x=1,y=2),b=c(3,4,list(z=5,s=6,t=list(m=7,n=8))))
j2= list(a=list(x=1,y=3),b=c(3,list(z=5,s=4,t=list(n=8))))
diffj_one(j1, j2)

}
