% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.r
\name{flattenj}
\alias{flattenj}
\title{Title transform a loaded json objects into a flatted data frame.}
\usage{
flattenj(dat)
}
\arguments{
\item{dat}{list. loaded result from a json file.}
}
\value{
data frame. flattening result.
}
\description{
Title transform a loaded json objects into a flatted data frame.
}
