% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_one.r
\name{flattenj_one}
\alias{flattenj_one}
\title{Title transform a loaded json objects into a flatted data frame.}
\usage{
flattenj_one(dat, seperator = ".")
}
\arguments{
\item{dat}{list. loaded result from a json file.}
}
\value{
data frame. flattening result.
}
\description{
Title transform a loaded json objects into a flatted data frame.
}
