% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.r
\name{loadj}
\alias{loadj}
\title{Title load a json file connection into an R list}
\usage{
loadj(file)
}
\arguments{
\item{file}{character. a json file connection}

\item{encoding}{character.Defaul to UTF-8. encoding method}
}
\value{
list. the loading result
}
\description{
Title load a json file connection into an R list
}
