% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.r
\name{descj}
\alias{descj}
\title{Title give the some description about the json list, such as keys and values.}
\usage{
descj(dat)
}
\arguments{
\item{dat}{list. a JSOn list.}
}
\value{
list. the result contains three parts. key_summary presents the description
of keys. value_summary presents the description of values. And stream_summary
presents the description of keys' direct father and children.
}
\description{
Title give the some description about the json list, such as keys and values.
}
