% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.r
\name{diffj}
\alias{diffj}
\title{Title calculating the difference between two jsons.}
\usage{
diffj(json_new, json_old)
}
\arguments{
\item{json_from}{list. The original json list.}

\item{json_to}{list. The target json list.}
}
\value{
data.frame. The difference between two json lists.
}
\description{
Title calculating the difference between two jsons.
}
