% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.r
\name{diffj}
\alias{diffj}
\title{Multiple JSON Objects Diff}
\usage{
diffj(json_new, json_old, sep = "@", primary)
}
\arguments{
\item{json_new}{list. The new json objects.}

\item{json_old}{list. The old json objects.}

\item{sep}{character. A character/string used to separate full keys in the nesting path.
Defaults to @ to avoid the occasional overriding. Not allowed to use some risky words like . and \.
When \code{compact=FALSE}, you need not to assign \code{sep} explicitly, unless @ has been used in the keys.}

\item{primary}{character. The primary key or path to identify a unique JSON object.
Usually a nesting path containing the \code{sep} character provided previously.}
}
\value{
list. Contains thw difference result, including create, delete and update information.
}
\description{
find the differences between multiple JSON objects including create, delete and update.
}
