% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.r
\name{diffj}
\alias{diffj}
\title{Multiple JSON Objects Diff}
\usage{
diffj(json_new, json_old, sep = "@", primary)
}
\arguments{
\item{json_new}{\code{list}. The new json objects.}

\item{json_old}{\code{list}. The old json objects.}

\item{sep}{\code{character}. A character/string used to separate keys in the nesting path.
Defaults to @ to avoid the occasional overriding. Not recommended to use some risky words like . and \.
When \code{compact=FALSE}, it is unnecessary to assign \code{sep} explicitly, unless @ has been used in the keys.}

\item{primary}{\code{character}. The primary key or path to identify a unique JSON object.
The value provided should contains the \code{sep} value to specify the nesting information unless it is an outermost key.}
}
\value{
\code{list}. Contains the difference result, including create, delete and update information.
}
\description{
Find the differences between multiple JSON objects yielded by create, delete and update operations.
}
\seealso{
\code{\link{diffj_one}}.
}
