% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_one.r
\name{diffj_one}
\alias{diffj_one}
\title{Single JSON Diff}
\usage{
diffj_one(json_new, json_old, sep = "@")
}
\arguments{
\item{json_new}{\code{list}. The new JSON list.}

\item{json_old}{\code{list}. The old JSON list.}

\item{sep}{\code{character}. A character/string used to separate keys in the nesting path.
Defaults to @ to avoid the occasional overriding. Not recommended to use some risky characters like . and \.}
}
\value{
\code{list}. The difference between two JSON objects. The \code{update} stores the data frame about the details of each type of change.
And the \code{summary} provides some useful summarization.
}
\description{
Calculate the difference between two JSON objects.
}
\details{
This function can find the difference between two JSON/list objects.
The difference information has three sources: path addition, path removal and value change.
All differences are reflected by '+(add)', and '-(delete)' chng_type in the result.
}
\examples{
library(mojson)
j1 = list(a=list(x=1,y=2),b=c(3,4,list(z=5,s=6,t=list(m=7,n=8))))
j2= list(a=list(x=1,y=3),b=c(3,list(z=5,s=4,t=list(n=8))))
diffj_one(j1, j2)

}
\seealso{
\code{\link{flattenj_one}}
}
